requirements.txt
absl-py==2.0.0
accelerate==0.23.0
ai4bharat-transliteration==1.1.3
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aksharamukha==1.9.7
alabaster==0.7.13
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
asteroid==0.6.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
bangla==0.0.2
bitarray==2.8.1
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
boltons==23.0.0
cached-property==1.5.2
cachetools==5.3.1
cdifflib==1.2.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
cmake==3.27.5
colorama==0.4.6
contourpy==1.1.1
coqpit==0.0.17
cryptography==41.0.4
ctranslate2==3.9.0
cycler==0.11.0
Cython==0.29.30
dateparser==1.1.8
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.18.1
editdistance==0.6.2
einops==0.6.1
encodec==0.1.1
exceptiongroup==1.1.3
fairseq==0.12.2
fastapi==0.103.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
ftfy==6.1.1
future==0.18.3
g2pkk==0.1.2
gevent==23.9.1
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
gradio==3.44.4
gradio_client==0.5.1
greenlet==2.0.2
grpcio==1.58.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.2
hydra-core==1.0.7
idna==3.4
imageio==2.31.3
imagesize==1.4.1
importlib-resources==6.1.0
indic-nlp-library @ file:///home/ruusaig/Projects/video-dubbing/IndicTrans2/indic_nlp_library
indic-numtowords==1.0.2
inflect==5.6.0
itsdangerous==2.1.2
jaconv==0.3.4
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
julius==0.2.7
k-diffusion==0.0.16
kiwisolver==1.4.5
kornia==0.7.0
langcodes==3.3.0
language-data==1.1
lazy_loader==0.3
librosa==0.10.0
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.3
marisa-trie==0.7.8
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mir-eval==0.7
mock==5.1.0
more-itertools==10.1.0
Morfessor==2.0.6
mosestokenizer==1.2.1
msgpack==1.0.6
multidict==6.0.4
multiprocess==0.70.15
nemo-text-processing==0.2.0rc0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.57.0
numpy==1.22.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.0.6
openai-whisper==20230918
openfile==0.0.7
orjson==3.9.7
packaging==23.1
pandas==2.0.3
pathos==0.3.1
pathtools==0.1.2
pb-bss-eval==0.0.2
pesq==0.0.4
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
portalocker==2.8.2
pox==0.3.3
ppft==1.7.6.7
progressbar2==4.2.0
protobuf==3.20.0
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyDeprecate==0.3.2
pydload==1.0.9
pydub==0.25.1
pyenchant==3.2.2
PyExecJS==1.5.1
Pygments==2.16.1
pykakasi==2.2.1
pynini==2.1.5
pynndescent==0.5.10
pyparsing==3.1.1
pypinyin==0.49.0
pysbd==0.3.4
pystoi==0.3.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-multipart==0.0.6
python-utils==3.7.0
pytorch-lightning==1.7.7
pytorch-ranger==0.1.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rpds-py==0.10.3
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soundfile==0.12.1
soxr==0.3.6
Sphinx==7.2.6
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2023.9.18
tiktoken==0.3.3
tokenizers==0.13.3
toolwrapper==2.1.0
toolz==0.12.0
torch==1.13.1
torch-optimizer==0.1.0
torch-stoi==0.1.2
torchaudio==0.13.1
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
torchvision==0.14.1
tqdm==4.66.1
trainer==0.0.31
trampoline==0.1.2
transformers==4.28.1
translators==5.8.5
triton==2.0.0
TTS==0.17.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uctools==1.3.0
ujson==5.8.0
umap-learn==0.5.1
Unidecode==1.3.6
urduhack==0.1.4
urllib3==1.26.16
uvicorn==0.23.2
wandb==0.15.10
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
wget==3.2
wrapt==1.15.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
